Renaming files and folders according to ARCH
