+++ LUFA/Drivers/Board/Board.h
@@ -163,7 +163,7 @@
 		#elif (BOARD == BOARD_XPLAINED_MINI)
 			#include "AVR8/XPLAINED_MINI/Board.h"
 		#elif (BOARD == BOARD_AVR64DU32_CNANO)
-			#include "AVR_Dx/AVR64DU32_CNANO/Board.h"
+			#include "AVRDX/AVR64DU32_CNANO/Board.h"
 		#else
 			#include "Board/Board.h"
 		#endif
