+++ LUFA/Drivers/Board/Buttons.h
@@ -161,7 +161,7 @@
 		#elif (BOARD == BOARD_U2S)
 			#include "AVR8/U2S/Buttons.h"
 		#elif (BOARD == BOARD_AVR64DU32_CNANO)
-			#include "AVR_Dx/AVR64DU32_CNANO/Buttons.h"
+			#include "AVRDX/AVR64DU32_CNANO/Buttons.h"
 		#else
 			#include "Board/Buttons.h"
 		#endif
