+++ LUFA/Drivers/Peripheral/SPI.h
@@ -69,7 +69,7 @@
 		#elif (ARCH == ARCH_XMEGA)
 			#include "XMEGA/SPI_XMEGA.h"
 		#elif (ARCH == ARCH_AVRDX)
-			#include "AVR_Dx/SPI_AVR_Dx.h"
+			#include "AVRDX/SPI_AVRDX.h"
 		#else
 			#error The SPI peripheral driver is not currently available for your selected architecture.
 		#endif
