+++ LUFA/Drivers/Peripheral/TWI.h
@@ -69,7 +69,7 @@
 		#elif (ARCH == ARCH_XMEGA)
 			#include "XMEGA/TWI_XMEGA.h"
 		#elif (ARCH == ARCH_AVRDX)
-			#include "AVR_Dx/TWI_AVR_Dx.h"
+			#include "AVRDX/TWI_AVRDX.h"
 		#else
 			#error The TWI peripheral driver is not currently available for your selected architecture.
 		#endif
