+++ LUFA/Drivers/USB/Core/USBController.h
@@ -153,7 +153,7 @@
 		#elif (ARCH == ARCH_XMEGA)
 			#include "XMEGA/USBController_XMEGA.h"
 		#elif (ARCH == ARCH_AVRDX)
-			#include "AVR_Dx/USBController_AVR_Dx.h"
+			#include "AVRDX/USBController_AVRDX.h"
 		#endif
 
 	/* Disable C linkage for C++ Compilers: */
