+++ LUFA/Drivers/USB/Core/USBInterrupt.h
@@ -63,7 +63,7 @@
 		#elif (ARCH == ARCH_XMEGA)
 			#include "XMEGA/USBInterrupt_XMEGA.h"
 		#elif (ARCH == ARCH_AVRDX)
-			#include "AVR_Dx/USBInterrupt_AVR_Dx.h"
+			#include "AVRDX/USBInterrupt_AVRDX.h"
 		#endif
 
 	/* Disable C linkage for C++ Compilers: */
